Add files via upload
Moving on from diagnosing the dispersal kernel, now that we know that the mean dispersal distance is how far the inoculum will travel proportional to the dispersal parameter, we now seek to generate a parameter space. This parameter space is important because it allows us to refine the input variables into a suitable model for early detection. Once the parameter space is evaluated, we will validate the early detection model by using it against simulation models of increasing complexity.